create interface file
